Extracts MockTimer into its own file
